space_launch/requirements.txt
anyio==3.6.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
spacexPython==1.0.1
sqlparse==0.4.3
urllib3==1.26.14
